refactor(fast_io): introduce IoBackend trait, eliminate stub duplication (#2244)
The non-Linux `io_uring_stub.rs` previously mirrored every public type and
function from `crate::io_uring` so cross-platform callers compiled without
`cfg`-gating. That redundancy was about 73 KB of mechanical duplication:
identical config structs, identical UAPI constants, identical error enums,
and identical `OpTag` / `SharedCompletion` definitions on both backends.

This commit pulls the plain-data types into a new `io_uring_common` module
that compiles on every target, and introduces an `IoBackend` trait that
both the live Linux backend (`LinuxIoUringBackend`) and the no-op stub
(`StubIoUringBackend`) implement. The Linux backend keeps only its
ring-construction behaviour; the stub keeps only the opaque handle types
whose methods always return `Unsupported`.

What moved into `io_uring_common`:

- `IoUringConfig` + `Default` + `for_large_files` / `for_small_files` /
  `allow_send_zc`
- `IoUringKernelInfo`, `SharedRingConfig`
- `OpTag` + `encode` / `decode`
- `SharedCompletion`
- `BufferRingConfig`, `BufferRingError` (with new `Unsupported` variant for
  the stub backend) and the shared `From<BufferRingError> for io::Error`
- `RegisteredBufferStats` + `miss_rate`, `RegisteredBufferStatus` +
  `is_*` helpers
- `buffer_id_from_cqe_flags` and the CQE bitmask constants
- All `IORING_OP_*`, `RENAME_*`, and `*_MIN_KERNEL` UAPI constants
- The `IoBackend` trait

Public API surface visible to other crates is unchanged: every name still
re-exports through `fast_io::*` at the existing paths.

Stub file before/after: 2246 -> 1805 lines (-441). The Linux io_uring
submodules collectively lost ~440 more lines that they no longer
re-declare. Net diff: 757 insertions, 1125 deletions for a 368-line
reduction once the new shared module is counted in.

Verified locally on macOS (the host running this build):

- `cargo check -p fast_io --no-default-features` (stub path, the path used
  by macos and windows non-iocp builds)
- `cargo check -p fast_io --all-features` (still selects the stub on
  macos; the Linux io_uring module is cfg-gated out)
- `cargo check --workspace --no-default-features`
- `cargo check --workspace --all-features`

CI will exercise the linux-gnu and linux-musl paths (where the real
Linux io_uring module is compiled) and run nextest on all three platforms.
