+++ crates/fast_io/src/io_uring/registered_buffers.rs
@@ -129,85 +129,7 @@ unsafe impl Send for RegisteredBufferGroup {}
 // is only accessed by the holder of a checked-out slot index.
 unsafe impl Sync for RegisteredBufferGroup {}
 
-/// Snapshot of [`RegisteredBufferGroup`] telemetry counters.
-///
-/// Returned by [`RegisteredBufferGroup::stats`]. Both fields are plain
-/// integers copied from atomic counters at the time of the call; under
-/// concurrent updates the snapshot is not strictly consistent across the
-/// two fields, but each individual value is accurate.
-///
-/// The miss rate is the fraction of acquire calls that returned `None`
-/// because every slot was in use - a forced fallback to non-registered
-/// `IORING_OP_READ` / `IORING_OP_WRITE`. A high miss rate is the primary
-/// signal driving the adaptive sizing design in
-/// `docs/audits/io-uring-adaptive-buffer-sizing.md`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct RegisteredBufferStats {
-    /// Number of `checkout` calls (whether they succeeded or not).
-    pub total_acquires: u64,
-    /// Number of `checkout` calls that returned `None`.
-    pub total_misses: u64,
-}
-
-impl RegisteredBufferStats {
-    /// Returns the miss rate as a fraction in `[0.0, 1.0]`.
-    ///
-    /// Returns `0.0` when no acquires have been recorded yet.
-    #[must_use]
-    pub fn miss_rate(&self) -> f64 {
-        if self.total_acquires == 0 {
-            return 0.0;
-        }
-        self.total_misses as f64 / self.total_acquires as f64
-    }
-}
-
-/// Owner-side state of fixed-buffer registration on an `io_uring` instance.
-///
-/// `Option<RegisteredBufferGroup>` alone cannot distinguish "the caller opted
-/// out via [`IoUringConfig::register_buffers`]" from "registration was
-/// attempted and the kernel rejected it" (low `RLIMIT_MEMLOCK`, seccomp,
-/// `IORING_MAX_REG_BUFFERS` exceeded, etc.). Both cases force the unfixed
-/// `IORING_OP_READ` / `IORING_OP_WRITE` fallback, but operators need to tell
-/// them apart when diagnosing a transfer that is unexpectedly slow.
-///
-/// Returned by [`IoUringReader::registered_buffer_status`] and
-/// [`IoUringWriter::registered_buffer_status`].
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum RegisteredBufferStatus {
-    /// Registration succeeded; `READ_FIXED` / `WRITE_FIXED` are in use.
-    Enabled,
-    /// The caller disabled registration via `IoUringConfig::register_buffers
-    /// = false`. No registration attempt was made.
-    Disabled,
-    /// Registration was attempted and rejected by the kernel. The owner
-    /// fell back silently to the unfixed opcodes. `reason` carries the
-    /// kernel's `errno` string for telemetry / tracing.
-    RegistrationFailed {
-        /// Reason returned by `IORING_REGISTER_BUFFERS` (formatted `errno`).
-        reason: String,
-    },
-}
-
-impl RegisteredBufferStatus {
-    /// True when `READ_FIXED` / `WRITE_FIXED` opcodes are active.
-    #[must_use]
-    pub fn is_enabled(&self) -> bool {
-        matches!(self, Self::Enabled)
-    }
-
-    /// True when the caller explicitly disabled registration.
-    #[must_use]
-    pub fn is_disabled(&self) -> bool {
-        matches!(self, Self::Disabled)
-    }
-
-    /// True when registration was attempted but rejected by the kernel.
-    #[must_use]
-    pub fn is_registration_failed(&self) -> bool {
-        matches!(self, Self::RegistrationFailed { .. })
-    }
-}
+pub use crate::io_uring_common::{RegisteredBufferStats, RegisteredBufferStatus};
 
 /// A checked-out buffer slot from a [`RegisteredBufferGroup`].
 ///
