+++ crates/fast_io/src/io_uring/shared_ring.rs
@@ -80,103 +80,7 @@ use super::registered_buffers::RegisteredBufferGroup;
 /// `opcode::PollAdd` in v0.7.
 const IORING_OP_POLL_ADD: u8 = 6;
 
-/// SQE `user_data` tag identifying the source channel of a completion.
-///
-/// Stored in the high 8 bits of `user_data`. Each value corresponds to a
-/// distinct demux arm in [`SharedRing::reap`]. Adding new tags is a backward-
-/// compatible change as long as existing values are preserved.
-#[repr(u8)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum OpTag {
-    /// File-side `IORING_OP_READ` or `IORING_OP_READ_FIXED` completion.
-    Read = 1,
-    /// File-side `IORING_OP_WRITE` or `IORING_OP_WRITE_FIXED` completion.
-    Write = 2,
-    /// Socket-side `IORING_OP_SEND` completion.
-    Send = 3,
-    /// Write-readiness probe (`IORING_OP_POLL_ADD` with `POLLOUT`).
-    PollWrite = 4,
-}
-
-impl OpTag {
-    /// Encodes the tag and a 56-bit op id into the SQE `user_data` field.
-    #[inline]
-    #[must_use]
-    pub fn encode(self, op_id: u64) -> u64 {
-        debug_assert!(op_id < (1 << 56), "op_id {op_id} overflows 56 bits");
-        ((self as u64) << 56) | (op_id & ((1u64 << 56) - 1))
-    }
-
-    /// Decodes a CQE `user_data` field into the source tag and op id.
-    ///
-    /// Returns `None` when the high 8 bits do not match a known tag value;
-    /// the caller should treat that as a corrupted completion.
-    #[inline]
-    pub fn decode(user_data: u64) -> Option<(Self, u64)> {
-        let tag = (user_data >> 56) as u8;
-        let op_id = user_data & ((1u64 << 56) - 1);
-        let parsed = match tag {
-            1 => Self::Read,
-            2 => Self::Write,
-            3 => Self::Send,
-            4 => Self::PollWrite,
-            _ => return None,
-        };
-        Some((parsed, op_id))
-    }
-}
-
-/// Demultiplexed result of one CQE drained from the shared ring.
-///
-/// Returned by [`SharedRing::reap`] for each completion the caller pulls off
-/// the CQ. The caller decides what to do with each variant - typically the
-/// reader path forwards `Read` results to its block matcher, and the writer
-/// path uses `PollWrite` to know when to follow up with a `Send`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum SharedCompletion {
-    /// File read completed; bytes read on success or `-errno` on failure.
-    Read {
-        /// Caller-supplied op id passed at submission.
-        op_id: u64,
-        /// Raw CQE result (bytes on success, negative errno on failure).
-        result: i32,
-    },
-    /// File write completed.
-    Write {
-        /// Caller-supplied op id passed at submission.
-        op_id: u64,
-        /// Raw CQE result (bytes on success, negative errno on failure).
-        result: i32,
-    },
-    /// Socket send completed.
-    Send {
-        /// Caller-supplied op id passed at submission.
-        op_id: u64,
-        /// Raw CQE result (bytes on success, negative errno on failure).
-        result: i32,
-    },
-    /// Write-readiness signalled by the kernel for the registered writer fd.
-    ///
-    /// The `revents` field carries the `POLLOUT`/`POLLERR`/`POLLHUP` bits
-    /// returned by the kernel; the caller should inspect them before issuing
-    /// the follow-up send.
-    PollWrite {
-        /// Caller-supplied op id passed at submission.
-        op_id: u64,
-        /// `revents` bitmask (raw poll flags from the kernel).
-        revents: i16,
-    },
-}
-
-/// Configuration for a [`SharedRing`].
-///
-/// Defaults match [`IoUringConfig::default`] for parity with the per-channel
-/// rings; callers tuning a session should override via the field below.
-#[derive(Debug, Clone, Default)]
-pub struct SharedRingConfig {
-    /// Backing io_uring configuration (SQ depth, registered buffer count, ...).
-    pub ring: IoUringConfig,
-}
+pub use crate::io_uring_common::{OpTag, SharedCompletion, SharedRingConfig};
 
 /// One io_uring ring shared by a reader and a writer fd in the same session.
 ///
